Add Lombok and delegate RoutingService to Graph
