+++ pom.xml
@@ -806,5 +806,12 @@
             <artifactId>javacsv</artifactId>
             <version>2.0</version>
         </dependency>
+        <!-- Lombok -->
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>1.18.12</version>
+            <scope>provided</scope>
+        </dependency>
     </dependencies>
 </project>
