+++ src/main/java/org/opentripplanner/routing/RoutingService.java
@@ -1,52 +1,32 @@
 package org.opentripplanner.routing;
 
-import com.google.common.collect.Multimap;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import graphql.ExecutionResult;
 import graphql.GraphQL;
 import graphql.execution.ExecutorServiceExecutionStrategy;
-import org.locationtech.jts.geom.Envelope;
+import lombok.experimental.Delegate;
 import org.opentripplanner.index.IndexGraphQLSchema;
 import org.opentripplanner.index.model.StopTimesInPattern;
 import org.opentripplanner.index.model.TripTimeShort;
-import org.opentripplanner.model.Agency;
-import org.opentripplanner.model.FeedInfo;
-import org.opentripplanner.model.FeedScopedId;
-import org.opentripplanner.model.MultiModalStation;
-import org.opentripplanner.model.Notice;
-import org.opentripplanner.model.Operator;
-import org.opentripplanner.model.Route;
-import org.opentripplanner.model.Station;
 import org.opentripplanner.model.Stop;
 import org.opentripplanner.model.Timetable;
 import org.opentripplanner.model.TimetableSnapshot;
-import org.opentripplanner.model.TransitEntity;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.TripPattern;
-import org.opentripplanner.model.calendar.CalendarService;
 import org.opentripplanner.model.calendar.ServiceDate;
 import org.opentripplanner.model.routing.RoutingResponse;
 import org.opentripplanner.routing.algorithm.RoutingWorker;
-import org.opentripplanner.routing.bike_rental.BikeRentalStationService;
 import org.opentripplanner.routing.core.RoutingRequest;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.routing.services.AlertPatchService;
-import org.opentripplanner.routing.vertextype.TransitStopVertex;
 import org.opentripplanner.standalone.server.Router;
 import org.opentripplanner.util.HttpToGraphQLMapper;
 
 import javax.ws.rs.core.Response;
-import java.util.BitSet;
 import java.util.Calendar;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.TimeZone;
 import java.util.concurrent.Executors;
-import java.util.stream.Collectors;
 
 /**
  * This is the entry point of all API requests towards the OTP graph. A new instance of this class
@@ -55,6 +35,7 @@ import java.util.stream.Collectors;
  */
 public class RoutingService {
 
+  @Delegate(types=Graph.class)
   private final Graph graph;
 
   /**
@@ -79,7 +60,7 @@ public class RoutingService {
    * Lazy-initialization of TimetableSnapshot
    * @return The same TimetableSnapshot is returned throughout the lifecycle of this object.
    */
-  private TimetableSnapshot getTimetableSnapshot() {
+  private TimetableSnapshot lazyGetTimeTableSnapShot() {
     if (this.timetableSnapshot == null) {
       timetableSnapshot = graph.getTimetableSnapshot();
     }
@@ -96,54 +77,6 @@ public class RoutingService {
       double lat, double lon, int radius
   ) {
     return StopFinder.findClosestStopsByWalking(graph, lat, lon, radius);
-  }
-
-  public Map<String, Map<String, Agency>> getAgenciesForFeedId() {
-    return graph.index.getAgenciesForFeedId();
-  }
-
-  public Map<FeedScopedId, Operator> getOperatorForId() {
-    return graph.index.getOperatorForId();
-  }
-
-  public Map<String, FeedInfo> getFeedInfoForId() {
-    return graph.index.getFeedInfoForId();
-  }
-
-  public Map<FeedScopedId, Stop> getStopForId() {
-    return graph.index.getStopForId();
-  }
-
-  public Map<FeedScopedId, Trip> getTripForId() {
-    return graph.index.getTripForId();
-  }
-
-  public Map<FeedScopedId, Route> getRouteForId() {
-    return graph.index.getRouteForId();
-  }
-
-  public Map<Stop, TransitStopVertex> getStopVertexForStop() {
-    return graph.index.getStopVertexForStop();
-  }
-
-  public Map<Trip, TripPattern> getPatternForTrip() {
-    return graph.index.getPatternForTrip();
-  }
-
-  public Multimap<Route, TripPattern> getPatternsForRoute() {
-    return graph.index.getPatternsForRoute();
-  }
-
-  public Multimap<Stop, TripPattern> getPatternsForStop() {
-    return graph.index.getPatternsForStop();
-  }
-
-  public Map<Station, MultiModalStation> getMultiModalStationForStations() {
-    return graph.index.getMultiModalStationForStations();
-  }
-
-  public Set<Route> getRoutesForStop(Stop stop) {
-    return graph.index.getRoutesForStop(stop);
   }
 
   /**
@@ -166,7 +99,7 @@ public class RoutingService {
   ) {
     return StopTimesHelper.stopTimesForStop(
         this,
-        getTimetableSnapshot(),
+        lazyGetTimeTableSnapShot(),
         stop,
         startTime,
         timeRange,
@@ -187,7 +120,7 @@ public class RoutingService {
   ) {
     return StopTimesHelper.getStopTimesForStop(
         this,
-        getTimetableSnapshot(),
+        lazyGetTimeTableSnapShot(),
         stop,
         serviceDate,
         omitNonPickups
@@ -201,7 +134,7 @@ public class RoutingService {
   public Collection<TripPattern> getPatternsForStop(Stop stop, boolean includeRealtimeUpdates) {
     return graph.index.getPatternsForStop(
         stop,
-        includeRealtimeUpdates ? getTimetableSnapshot() : null);
+        includeRealtimeUpdates ? lazyGetTimeTableSnapShot() : null);
   }
 
   /**
@@ -210,7 +143,7 @@ public class RoutingService {
    * without making a fake routing request.
    */
   public Timetable getTimetableForTripPattern(TripPattern tripPattern) {
-    TimetableSnapshot timetableSnapshot = getTimetableSnapshot();
+    TimetableSnapshot timetableSnapshot = lazyGetTimeTableSnapShot();
     return timetableSnapshot != null ?
         timetableSnapshot.resolve(tripPattern, new ServiceDate(Calendar.getInstance().getTime()))
         : tripPattern.scheduledTimetable;
@@ -223,69 +156,8 @@ public class RoutingService {
     return HttpToGraphQLMapper.mapExecutionResultToHttpResponse(executionResult);
   }
 
-  /**
-   * Fetch an agency by its string ID, ignoring the fact that this ID should be scoped by a feedId.
-   * This is a stopgap (i.e. hack) method for fetching agencies where no feed scope is available. I
-   * am creating this method only to allow merging pull request #2032 which adds GraphQL. Note that
-   * if the same agency ID is defined in several feeds, this will return one of them at random. That
-   * is obviously not the right behavior. The problem is that agencies are not currently keyed on an
-   * FeedScopedId object, but on separate feedId and id Strings. A real fix will involve replacing
-   * or heavily modifying the OBA GTFS loader, which is now possible since we have forked it.
-   */
-  public Agency getAgencyWithoutFeedId(String agencyId) {
-    // Iterate over the agency map for each feed.
-    for (Map<String, Agency> agencyForId : getAgenciesForFeedId().values()) {
-      Agency agency = agencyForId.get(agencyId);
-      if (agency != null) {
-        return agency;
-      }
-    }
-    return null;
-  }
-
-  /**
-   * Construct a set of all Agencies in this graph, spanning across all feed IDs. I am creating this
-   * method only to allow merging pull request #2032 which adds GraphQL. This should probably be
-   * done some other way, see javadoc on getAgencyWithoutFeedId.
-   */
-  public Set<Agency> getAllAgencies() {
-    Set<Agency> allAgencies = new HashSet<>();
-    for (Map<String, Agency> agencyForId : getAgenciesForFeedId().values()) {
-      allAgencies.addAll(agencyForId.values());
-    }
-    return allAgencies;
-  }
-
-  public Collection<Notice> getNoticesByEntity(TransitEntity<?> entity) {
-    Collection<Notice> res = graph.getNoticesByElement().get(entity);
-    return res == null ? Collections.emptyList() : res;
-  }
-
-  /**
-   * Get a list of all operators spanning across all feeds.
-   */
-  public Collection<Operator> getAllOperators() {
-    return getOperatorForId().values();
-  }
-
-  public BitSet getServicesRunningForDate(ServiceDate date) {
-    return graph.servicesRunning(date);
-  }
-
-  public TripPattern getTripPatternForId(String id) {
-    return graph.tripPatternForId.get(id);
-  }
-
-  public Collection<TripPattern> getTripPatterns() {
-    return graph.tripPatternForId.values();
-  }
-
-  public Collection<Notice> getNotices() {
-    return graph.getNoticesByElement().values();
-  }
-
   public List<TripTimeShort> getStopTimesForTripAndDate(Trip trip, ServiceDate serviceDate) {
-    TimetableSnapshot timetableSnapshot = getTimetableSnapshot();
+    TimetableSnapshot timetableSnapshot = lazyGetTimeTableSnapShot();
     return timetableSnapshot != null
         ? TripTimeShort.fromTripTimes(
             timetableSnapshot.resolve(getPatternForTrip().get(trip), serviceDate),
@@ -294,46 +166,10 @@ public class RoutingService {
         : null;
   }
 
-  public Collection<Stop> getStopsByBoundingBox(Envelope envelope) {
-    return graph.streetIndex
-        .getTransitStopForEnvelope(envelope)
-        .stream()
-        .map(TransitStopVertex::getStop)
-        .collect(Collectors.toList());
-  }
-
-  public Station getStationById(FeedScopedId id) {
-    return graph.stationById.get(id);
-  }
-
-  public Collection<Station> getStations() {
-    return graph.stationById.values();
-  }
-
-  public CalendarService getCalendarService() {
-    return graph.getCalendarService();
-  }
-
-  public TimeZone getTimeZone() {
-    return graph.getTimeZone();
-  }
-
-  public BikeRentalStationService getBikerentalStationService() {
-    return graph.getService(BikeRentalStationService.class);
-  }
-
-  public AlertPatchService getSiriAlertPatchService() {
-    return graph.getSiriAlertPatchService();
-  }
-
-  public MultiModalStation getMultiModalStationById(FeedScopedId feedScopedId) {
-    return graph.multiModalStationById.get(feedScopedId);
-  }
-
   public List<TripTimeShort> getTripTimesShort(Trip trip, ServiceDate serviceDate) {
     return TripTimesShortHelper.getTripTimesShort(
         this,
-        getTimetableSnapshot(),
+        lazyGetTimeTableSnapShot(),
         trip,
         serviceDate
     );
